Move widget class to different package
